mondrian: perf: Disable Display PID Affine feature
This feature dynamically affines SF and HWC to the foreground CPU set
(or big cluster), in some scenarios to improve performance, and then
back to system-background CPU set (little cluster).
However, we already affine SF and HWC to the foreground CPU set in
platform, so let's just keep it there always rather than having perfd
move it to system-background.